% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-mci.R
\name{read_mci}
\alias{read_mci}
\title{Reads 'mcI' files}
\usage{
read_mci(mci_file)
}
\arguments{
\item{mci_file}{An mcI file}
}
\value{

}
\description{
Reads 'mcI' files
}
\examples{
file <- "ftp://dfoftp.ocean.dal.ca/pub/dfo/BSRTO/2018-2019/mcI/18082923.mcI"
read_mci(file)

}
